requirements.txt
bandit==1.7.4
coverage==6.3.2
flake8==4.0.1
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
stringcase==1.2.0
yahoo-oauth==1.1
